email_daily_news/pyproject.toml
[tool.poetry]
name = "email-daily-news"
version = "0.1.0"
description = "Email News Articles to Yourself Daily"
authors = ["Brad Penney <penney.brad@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
datetime = "^5.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
